LPS-90290 Move to rest module
It makes more sense there
